+++ neovim/lua/mappings.lua
@@ -1,53 +1,4 @@
-function map(mode, shortcut, command, noremap, unique)
-  vim.api.nvim_set_keymap(mode, shortcut, command,
-    { noremap = noremap, unique = unique, silent = true }
-  )
-end
-
-function unmap(mode, shortcut)
-  vim.api.nvim_del_keymap(mode, shortcut)
-end
-
-function buf_map(bufnr, mode, shortcut, command, noremap, unique)
-  vim.api.nvim_buf_set_keymap(bufnr, mode, shortcut, command,
-    { noremap = noremap, unique=unique, silent = true }
-  )
-end
-
-function buf_unmap(bufnr, mode, shortcut)
-  vim.api.nvim_buf_del_keymap(bufnr, mode, shortcut)
-end
-
-
--- Unique nmap
-function nmap_u(shortcut, command)
-  map('n', shortcut, command, false, true)
-end
-
--- Non unique nmap
-function nmap_nu(shortcut, command)
-  map('n', shortcut, command, false, false)
-end
-
--- Unique nnoremap
-function nnoremap_u(shortcut, command)
-  map('n', shortcut, command, true, true)
-end
-
--- Non unique nnoremap
-function nnoremap_nu(shortcut, command)
-  map('n', shortcut, command, true, false)
-end
-
--- Unique nnoremap
-function buf_nnoremap_u(bufnr, shortcut, command)
-  buf_map(bufnr, 'n', shortcut, command, true, true)
-end
-
--- Non unique nnoremap
-function buf_nnoremap_nu(bufnr, shortcut, command)
-  buf_map(bufnr, 'n', shortcut, command, true, false)
-end
+require("utils")
 
 -- Load nvim settings
 -- =============================================
@@ -113,26 +64,3 @@ nnoremap_u("<leader>tt", ":tabedit <cr>") -- Open new tab
 nnoremap_nu("<Esc>", "<cmd>nohlsearch<CR>")
 -- =============================================
 
-
--- Telescope mappings
--- =============================================
-nnoremap_u("<leader>ff", "<cmd>lua require('telescope.builtin').find_files()<cr>")
-nnoremap_u("<leader>fs", "<cmd>lua require('telescope.builtin').grep_string()<cr>")
-nnoremap_u("<leader>fl", "<cmd>lua require('telescope.builtin').live_grep()<cr>")
-nnoremap_u("<leader>fb", "<cmd>lua require('telescope.builtin').buffers()<cr>")
-nnoremap_u("<leader>fh", "<cmd>lua require('telescope.builtin').help_tags()<cr>")
-nnoremap_u("<leader>fo", "<cmd>lua require('telescope.builtin').oldfiles()<cr>")
-nnoremap_u("<leader>fc", "<cmd>lua require('telescope.builtin').command_history()<cr>")
-nnoremap_u("<leader>fgs", "<cmd>lua require('telescope.builtin').git_status()<cr>")
--- =============================================
-
-
--- Colorscheme setup
--- Ideally this is done post install, but keeping it here for now.
--- =============================================
-vim.cmd([[
-  let g:sonokai_style = 'andromeda'
-  let g:sonokai_better_performance = 1
-  colorscheme sonokai
-]])
--- =============================================
