+++ neovim/lua/plugins/catppuccin.lua
@@ -1,0 +1,9 @@
+return {
+  "catppuccin/nvim",
+  name = "catppuccin",
+  priority = 1000,
+  config = function()
+    -- catppuccin-latte, catppuccin-frappe, catppuccin-macchiato, catppuccin-mocha
+    vim.cmd([[colorscheme catppuccin-mocha]])
+  end,
+}
